Remove Npc class from the main program. Update the project with an NPC.cs
